requirements.txt
feedparser>=6.0.11
requests>=2.31.0
google-genai>=1.2.0
python-telegram-bot[job-queue]>=21.0
python-dotenv>=1.0.1
schedule>=1.2.2
